Rename IMetricUpdateCallback to IOnPollMetricsCallback (#21)
The current name implies that the metric producer should push its
metrics. The intention is to give it a chance to update its metrics
because they are about to be polled.

The interface was also updated to let the implementation know which
members will be polled. This should facilitate the most common pattern
which is to check if a value needs to be updated (e.g. make a
measurement), and then updating it if needed.

Since the input is constructed as a linq expression, the computation of
polled values should be quite cheap if the producer does not care.